Add file share target
File share target allows Blaze to be shown in share trays and
receive files once installed on the device.
Selected files are queued to a central context and can be shared
with other peers after joining a room. FAB is updated to show
'Send selected file(s)' so that the files can be sent at any point
of time.

Fixes #47