Add eager socket aliveness checks when user with same name joins same room
If a user A is trying to join a room which already has user B with
the same name as A, Blaze first checks if user B is "alive" by
sending PING message and anticipating a PONG within some
timeframe.
If a PONG is received by user B, user A is not let into the room
because of the name clash.
However, if the PONG is not received by user B within the alotted
timeframe, it is assumed that user B has disconnected and the
stray socket on server is closed. User A is also let in because
there is no name clash in this scenario.

Fixes #84